requirements-healthcare.txt
# ── Core ──────────────────────────────────────────────────────────────────────
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
alembic>=1.12.0

# ── Auth / Security ────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pyotp>=2.9.0
qrcode[pil]>=7.4.0

# ── AI / LLM ───────────────────────────────────────────────────────────────────
openai>=1.0.0

# ── Task Queue (24/7 Workflows) ────────────────────────────────────────────────
celery>=5.3.0
redis>=5.0.0

# ── Payments ──────────────────────────────────────────────────────────────────
stripe>=7.0.0

# ── Notifications ──────────────────────────────────────────────────────────────
twilio>=8.0.0
sendgrid>=6.10.0

# ── PDF Generation ─────────────────────────────────────────────────────────────
reportlab>=4.0.0

# ── Frontend Templating ────────────────────────────────────────────────────────
jinja2>=3.1.0
aiofiles>=23.1.0

# ── Scheduling ────────────────────────────────────────────────────────────────
apscheduler>=3.10.0

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx>=0.25.0
requests>=2.31.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
